aoe-indicators: refactor plugin name & move AoE projectile
