+++ pom.xml
@@ -45,7 +45,7 @@
         <glslang.path></glslang.path>
 
 		<rs.version>189</rs.version>
-		<sanlite.version>1.13.4</sanlite.version>
+		<sanlite.version>1.14.0</sanlite.version>
 	</properties>
 
 	<licenses>
