+++ runelite-client/src/main/java/net/runelite/client/plugins/areaofeffectindicators/AreaOfEffectProjectile.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.game;
+package net.runelite.client.plugins.areaofeffectindicators;
 
 import lombok.Data;
 import net.runelite.api.Actor;
